E6C1
Changed the way the ellipse around the brightest pixel is drawn so it covers the entire flashlight spot (The program picks up the brightest pixel at the left, and since all the other pixels are as bright, they don't get picked up. Therefore, the shape is usually drawn at the left of the light source. However, I want the bouncers to react to all the brightest pixels - not just the left one. For now, my solution is to adjust the shape so it is drawn 50 pixels to the right)