E6C3
Changed from finding the brightest pixel to the reddest pixel. First, because I need to do that later in my project anyways. Secondly, because using a light to control the target created a lens flare that was bright enough to be detected and used instead of the light as the brightest pixel.