E6C5
Changed the code so that when a bouncer touches a paddle, it changes to its color (red or blue). They are not supposed to bounce off the paddles. (In fact, I think "paintbrush' would be a more accurate word to describe the paddles at this point).

There is also a score system that keeps track of how many bouncers are of each color (red or blue)