E10C1
added a rotating paddle class (to make a paddle consisting of an array or circles). tried to use rotate() to move it, but it didnt work because the rotate() function resets at the end of draw(). So i created a pvector, set it to the paddle's position, rotated it. then set the paddle position to the pvector. Next step: getting several balls in a row...