C7
RealBricks: Changed the ball interacion with the redPaddle. Before, the ball would get stuck in the paddle and then go down instead of up. That was because the paddle would move over the ball sometimes, and the changeVY function would be triggered repeatedly. Now, the ball checks if the ball is going down. If it is, it switches the VY so it bounces. If the ball is already going up, it does nothing.