C8
added sounds when the ball hits paddle or walls.

tweaked Head Meets Brick. Now, the ball bounces off the paddle and walls (because of an if/else thing in the ball class), the paddle doesn't have a huge stroke around it (added noStroke to paddle display), and the bricks are aligned properly (by switching image mode to CORNER)